Put TLSIdentity tests into it's own test class.
